% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fizzbuzz.R
\name{fizzbuzz}
\alias{fizzbuzz}
\title{Fizzbuzz function}
\usage{
fizzbuzz(x)
}
\arguments{
\item{x}{\emph{Numeric} vector to use for the fizzbuzz function}
}
\value{
Returns a \emph{character} vector of values, "fizz", "buzz", or "fizzbuzz"
}
\description{
A short description...
}
\examples{
fizzbuzz(1:10)
fizzbuzz(10:1)

}
